requirements.txt
# Core dependencies for Hallo2 video generation
PyYAML>=6.0.1

# HuggingFace Hub for model downloads
# Must be compatible with transformers requirement: >=0.19.3,<1.0
huggingface-hub>=0.19.3,<1.0

# Hallo2 will install its own dependencies when cloned
# This file is for standalone utilities only
